requirements.txt
# requirements.txt
cryptography>=3.4.8
requests>=2.25.1
pandas>=1.3.0
openpyxl>=3.0.7 # Для работы с .xlsx, если нужно
pyyaml>=5.4.1
# Для GUI (выберите один, если будете делать GUI)
PyQt6>=6.2.0
# или
# tkinter (обычно встроен в Python)